Update field-resolvers.md
This issue is tricky. When you want to resolve a Date field for example, you'll be tempted to write 
```
type:Date,
resolveAs:{
type: Date,
resolver() {...}
```
the first type expects a Constructor, the second a GraphQL type name as a string, so one can easily forget the quotes. The error message is then hardly understandable. In this example the type name will become "function Date()" and create invalid gql schema, but the solution is not trivial. Sso that would be nice to add this on the doc (or improve the error message).