Deployment docs update
The Meteor Up has a webpage with really great documentation, let's utilise that and direct to it whenever possible. This eliminates some possibilities for outdated information and clarifies where to look for more information if needed.

Added a link to a guide from DigitalOcean on an initial server setup. It might be useful for those who haven't done it before or need a refresher.

Updated the example conf with a configuration that I know works on an Ubuntu 16.04 dev machine to Ubuntu 16.04 server. It should just work though.

Refactored some sentences to remove redundant statements, clarify more specifically
Use some sentences directly from the official MUP docs when I couldn't come up with a proper explanation myself. 

The mup commands have been added to the docs. They are directly copied from the official docs and should be included for ease-of-use and clarification.